Renamed Algotithms Solutions
