+++ Algorithms/Strings/CamelCase/CameCase_Solution.java
@@ -2,7 +2,7 @@
 
 import java.util.Scanner;
 
-public class Solution {
+public class CameCase_Solution {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
         String str = scan.next();
